requirements.txt
django>=5.1.5
django-allauth>=65.3.1
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.4.0
drf-spectacular>=0.28.0
pillow>=11.1.0
whitenoise>=6.8.2